requirements.txt
beautifulsoup4
requests
openpyxl
pillow